LogPage: Unstub $wgLang before use
The code using $wgLang is also moved inside a conditional, to avoid
unstubbing $wgLang when unnecessary.

A possible improvement is to use Skin::getLanguage() instead of $wgLang,
but that's not guaranteed to be safe, and we want to unblock the train
without introducing new issues.

Bug: T245178
Change-Id: Ifb06d9b890456092633e38b14674cc3ef48690e3
(cherry picked from commit 1967ba0ad6db90788fd86aafe34689d1b0173eb7)