maintenance: Remove sql.php temporarily
It's dangerously broken.

This is supposed to be a read-only ad-hoc debugging script but in actually
triggers 'LoadExtensionSchemaUpdates', which is effectively update.php, but
without any of the safe guards, and at a time where you least expect write
queries to happen, let alone administrative queries that drop tables and
run other schema migrations.

Bug: T157651
Depends-On: I8b0a3a6e096c50996b166a0b5f7159a29bf77deb
Change-Id: I8d18a7f5b1731333aa40a9d04dafdb7176cf1d52