RevisionUndeleted hook: Add $oldPageId
When the hook was first introduced, I misunderstood that the row's
ar_page_id was not always the same as the newly restored revision's
page, because the page id could change following undeletion.

This commit added a second parameter, the old page id, to the hook,
a parameter that was present in the ArticleRevisionUndeleted hook that
RevisionUndeleted replaced.

Since both Flow and LiquidThreads were updated to use the new hook
before the issue with the page id was discovered, cherry picking to
the current train in case issues emerge due to the misunderstanding.
Both Flow and LiquidThreads have been updated to use the second parameter
now provided, but those changes have not been backported (yet), though
cherry picks will be available in case issues arise and the fixes should
be merge to the deployed versions.

Bug: T248727
Change-Id: I4e3ea4519d34de7c67ad8999575affe77d756f08
(cherry picked from commit 5127acb8ec226baebe4864ace190c4264b2f0023)