Improve logging for "main slot of revision not found in database"
The "main slot of revision not found in database" issue described
in T212428 has been ongoing for over a year and is throwing an
exception that is difficult to filter out during logstash triage.
Remove the revision id from the exception message, making it easier
to filter out. Instead, include this in new diagnostic logging.

Bug: T256127
Change-Id: I8f986daeab44c0b1c6b6e3fbf126b32053bd58e0